Renamed Loginfile to Log, and changed to singleton Log.java is now a enum-based singleton Changed calls in PassengerQueue and CheckinCounter to account for this
