+++ CheckInTerminal/src/part2/Main.java
@@ -13,14 +13,13 @@ public class Main {
         AllFlights flights = csvReader.getFlights();
         SimTime t = new SimTime();
         Timer timer = new Timer(t);
-        Loginfile log = new Loginfile(timer);
-        CheckinCounter c1 = new CheckinCounter(1,flights,t,timer,log);
+        CheckinCounter c1 = new CheckinCounter(1,flights,t,timer);
         counters.add(c1); 
-        CheckinCounter c2 = new CheckinCounter(2,flights,t,timer,log);
+        CheckinCounter c2 = new CheckinCounter(2,flights,t,timer);
         counters.add(c2);
-        CheckinCounter c3 = new CheckinCounter(3,flights,t,timer,log);
+        CheckinCounter c3 = new CheckinCounter(3,flights,t,timer);
         counters.add(c3);
-        PassengerQueue pq = new PassengerQueue(timer,log);
+        PassengerQueue pq = new PassengerQueue(timer);
         ManagementGUI gui = new ManagementGUI(timer,t,counters,flights.getAllFlights());
         
         pq.registerObserver(gui);
@@ -44,6 +43,7 @@ public class Main {
         c3.start();
         Runtime.getRuntime().addShutdownHook(new Thread(new Runnable() {
 	        public void run() {
+                Log log = Log.INSTANCE;
                 log.savefile();
 	        }
 	    }, "Shutdown-thread"));
